Fix api url error, and create component for pagintor.
