+++ club/api/urls.py
@@ -19,7 +19,7 @@ urlpatterns = [
     path('reserve/delete', ReserveDelete.as_view()),
     path('series/<page_number>', SeriesListView.as_view()),
     path('series/genre/<series_genre>/<page_number>', SeriesByGenreView.as_view()),
-    path('series/<series_title>', SeriesByTitleDetailView.as_view()),
+    path('series/detail/<series_title>', SeriesByTitleDetailView.as_view()),
     path('shared/<club_id>/<page_number>', SharedAccessListView.as_view()),
     path('request_shared_access', SharedAccessRequest.as_view()),
     path('access_request_respond', SharedAccessRequestResponse.as_view())
